# api-gateway/swagger.yaml
openapi: 3.0.1
info:
  title: book-recommendations
  description: book-recommendations
  version: v1
  contact:
    name: Julio Scheidt
    email: julio@blackdevs.com.br
servers:
  - url: http://book-recommendations.{NAMESPACE}.svc.cluster.local:3080/
...
paths:
...
  /v1/auth/signup:
    post:
      tags:
        - AuthServices
      summary: user signup
      operationId: userSignUp
      requestBody:
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/UserBody'
                - type: object
                  properties:
                    password:
                      type: string
      responses:
        '201':
          description: 'Created'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSignUpSuccessResponse'
        '400':
          description: 'Bad Request'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultResponseBadRequest'
        '500':
          description: 'Internal Server Error'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultResponseServerError'
  /v1/auth/signin:
    post:
      tags:
        - AuthServices
      summary: user signin
      operationId: userSignIn
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: 'OK'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSignInSuccessResponse'
        '400':
          description: 'Bad Request'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultResponseBadRequest'
        '500':
          description: 'Internal Server Error'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultResponseServerError'
...
  /v1/user:
    post:
      tags:
        - UserServices
      summary: create user
      operationId: createUser
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/UserBody'
                - type: object
                  properties:
                    password:
                      type: string
      responses:
        '201':
          description: 'Created'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateUserSuccessResponse'
        '400':
          description: 'Bad Request'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultResponseBadRequest'
        '401':
          description: 'Unauthorized'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultResponseUnauthorized'
        '500':
          description: 'Internal Server Error'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultResponseServerError'

    get:
      tags:
        - UserServices
      summary: list user
      operationId: listUser
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          description: Page to search
          required: true
          schema:
            type: integer
            example: 0
            minimum: 0
        - name: size
          in: query
          description: Size of page
          required: true
          schema:
            type: integer
            example: 50
            maximum: 100
      responses:
        '200':
          description: 'OK'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListUserSuccessResponse'
        '401':
          description: 'Unauthorized'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultResponseUnauthorized'
        '500':
          description: 'Internal Server Error'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultResponseServerError'
  /v1/user/{uuid}:
    get:
      tags:
        - UserServices
      summary: get user
      operationId: getUser
      security:
        - BearerAuth: []
      parameters:
      - name: uuid
        in: path
        description: Uuid for user
        required: true
        schema:
          $ref: '#/components/schemas/Uuid'
      responses:
        '200':
          description: 'OK'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetUserSuccessResponse'
        '400':
          description: 'Bad Request'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultResponseBadRequest'
        '401':
          description: 'Unauthorized'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultResponseUnauthorized'
        '404':
          description: 'Not Found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultResponseNotFound'
        '500':
          description: 'Internal Server Error'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultResponseServerError'

    put:
      tags:
        - UserServices
      summary: update user
      operationId: updateUser
      security:
        - BearerAuth: []
      parameters:
      - name: uuid
        in: path
        description: Uuid for user
        required: true
        schema:
          $ref: '#/components/schemas/Uuid'
      requestBody:
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/UserBody'
                - type: object
                  properties:
                    password:
                      type: string
      responses:
        '202':
          description: 'Accepted'
        '400':
          description: 'Bad Request'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultResponseBadRequest'
        '401':
          description: 'Unauthorized'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultResponseUnauthorized'
        '404':
          description: 'Not Found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultResponseNotFound'
        '500':
          description: 'Internal Server Error'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultResponseServerError'

    delete:
      tags:
        - UserServices
      summary: delete user
      operationId: deleteUser
      security:
        - BearerAuth: []
      parameters:
      - name: uuid
        in: path
        description: Uuid for user
        required: true
        schema:
          $ref: '#/components/schemas/Uuid'
      responses:
        '204':
          description: 'No Content'
        '400':
          description: 'Bad Request'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultResponseBadRequest'
        '401':
          description: 'Unauthorized'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultResponseUnauthorized'
        '404':
          description: 'Not Found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultResponseNotFound'
        '500':
          description: 'Internal Server Error'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultResponseServerError'
...
  /v1/user/rate/{user_uuid}:
    get:
      tags:
        - UserRateServices
      summary: get user rate
      operationId: getUserRate
      security:
        - BearerAuth: []
      parameters:
      - name: user_uuid
        in: path
        description: Uuid for user
        required: true
        schema:
          $ref: '#/components/schemas/Uuid'
      responses:
        '200':
          description: 'OK'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetUserRateSuccessResponse'
        '400':
          description: 'Bad Request'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultResponseBadRequest'
        '401':
          description: 'Unauthorized'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultResponseUnauthorized'
        '404':
          description: 'Not Found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultResponseNotFound'
        '500':
          description: 'Internal Server Error'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultResponseServerError'

    delete:
      tags:
        - UserRateServices
      summary: delete user rate
      operationId: deleteUserRate
      security:
        - BearerAuth: []
      parameters:
      - name: user_uuid
        in: path
        description: Uuid for user
        required: true
        schema:
          $ref: '#/components/schemas/Uuid'
      responses:
        '204':
          description: 'No Content'
        '400':
          description: 'Bad Request'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultResponseBadRequest'
        '401':
          description: 'Unauthorized'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultResponseUnauthorized'
        '404':
          description: 'Not Found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultResponseNotFound'
        '500':
          description: 'Internal Server Error'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultResponseServerError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
  schemas:
    DefaultResponseBadRequest:
      type: object
      properties:
        message:
          type: string
          description: 'Status code'
          example: 'Bad Request'
    DefaultResponseUnauthorized:
      type: object
      properties:
        message:
          type: string
          description: 'Status code'
          example: 'Unauthorized'
    DefaultResponseNotFound:
      type: object
      properties:
        message:
          type: string
          description: 'Status code'
          example: 'Not Found'
    DefaultResponseServerError:
      type: object
      properties:
        message:
          type: string
          description: 'Status code'
          example: 'Internal Server Error'
    Uuid:
      type: string
      example: '17f182b0dfc002737d507f74'
    CreatedAt:
      type: string
      format: date-time
      example: '2021-01-01T12:00:00.000Z'
    UpdatedAt:
      type: string
      format: date-time
      example: '2021-01-01T12:00:00.000Z'
    UserBody:
      type: object
      properties:
        name:
          type: string
        surname:
          type: string
        email:
          type: string
        phone:
          type: string
          format: phone
          example: '1199999999'
    UserFull:
      allOf:
        - type: object
          properties:
            uuid:
              $ref: '#/components/schemas/Uuid'
            created_at:
              $ref: '#/components/schemas/CreatedAt'
            updated_at:
              $ref: '#/components/schemas/UpdatedAt'
        - $ref: '#/components/schemas/UserBody'
    UserRateBody:
      allOf:
        - type: object
          properties:
            user_uuid:
              $ref: '#/components/schemas/Uuid'
    UserRateFull:
      allOf:
        - $ref: '#/components/schemas/UserRateBody'
        - type: object
          properties:
            rates:
              type: array
              items:
                type: object
                properties:
                  book_uuid:
                    $ref: '#/components/schemas/Uuid'
                  rate:
                    type: number
            created_at:
              $ref: '#/components/schemas/CreatedAt'
            updated_at:
              $ref: '#/components/schemas/UpdatedAt'
...
    BearerToken:
      type: string
      example: >-
        eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1dWlkIjoiMTdmOWFiMzRmMDUwZmM0MTk5Y2Y5MThiIiwibmFtZSI6Imp1bGlvIiwic3VybmFtZSI6ImNlc2FyIiwiZW1haWwiOiJqdWxpb0BtYWlsLmNvbSIsInBob25lIjoiKzU1NDE5OTg4Nzc2NiIsImlhdCI6MTY0NzU2NzkxNiwiZXhwIjoxNjQ3NTcxNTE2LCJhdWQiOiJ1cm46Ym9vay1yZWNvbW1lbmRhdGlvbnM6YXBpLWdhdGV3YXkiLCJpc3MiOiJ1cm46Ym9vay1yZWNvbW1lbmRhdGlvbnM6dXNlcnMtbWljcm9zZXJ2aWNlIn0.Anff7tEPzaqIlpPFgqNNoxKg0CapOxLl6EaI5I3EmR48VyMJYsm3MWhRpXoTpoGoH_cUXu9AVXAY39kueVZP2-eJmq6sHh8awrjlEOh14mx9og_inaOZps8g-tUE0g93LAbPMAnxeMU584Gorg3xaJpe-RsJn5QttWNX2WgqmFJKKZ4v38Bb5w75SzLFQ0G5FiobwNHJmmqV5z4HFqkRyU-6D65sn_LwrS_TgQ8cw73bpHyfQfoDCkngNjGEpzvi6CEvWNkiPLAgHJiwrnJHNVIF1zt0Ltj3DxgRh_R4TEpLpmUTKxBpTjGbSzbuXSGYqdLEy0EFd9CinOyE04h_0Q
    UserSignUpSuccessResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/BearerToken'
    UserSignInSuccessResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/BearerToken'
    CreateUserSuccessResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/Uuid'
    GetUserSuccessResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/UserFull'
    ListUserSuccessResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/UserFull'
...
    GetUserRateSuccessResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/UserRateFull'
...